// furia-backend/src/users/dto/update-user.dto.ts
export class UpdateUserDto {
  address?: string;
  cpf?: string;
  interests?: string;
  activities?: string;
  events?: string;
  purchases?: string;
  twitter?: string;
  instagram?: string;
  facebook?: string;
  steam?: string;
  valorant?: string;
  gamersclub?: string;
  faceit?: string;
  lol?: string;
  fifa?: string;
}